convert ConfirmationInput to ts
