--- src/ConfirmationInput/index.d.ts
@@ -1,3 +1,0 @@
-import ConfirmationInput, { ConfirmationInputProps } from './ConfirmationInput';
-export default ConfirmationInput;
-export { ConfirmationInputProps };
