--- src/ConfirmationInput/index.js
@@ -1,2 +1,0 @@
-import ConfirmationInput from './ConfirmationInput';
-export default ConfirmationInput;
