+++ src/ConfirmationInput/index.ts
@@ -1,0 +1,3 @@
+import ConfirmationInput, { ConfirmationInputProps, Props } from './ConfirmationInput';
+export default ConfirmationInput;
+export { ConfirmationInputProps, Props };
